requirements/common.txt
Django>=2.1.6
djangorestframework==3.9.1
django-filter==2.1.0
psycopg2-binary==2.7.7
